abstract kubernetes clients
